completed the jwt authorization

+++ controllers/goalController.js
@@ -1,6 +1,8 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
 const User = require("../models/userModel");
+
+
 // @desc Get Goals
 // @route Get /api/goals
 // @access Private
@@ -9,6 +11,8 @@ const getGoals = asyncHandler(async (req, res) => {
 
   res.status(200).json(goals);
 });
+
+
 // @desc Post Goals
 // @route Post /api/goals
 // @access Private
@@ -29,6 +33,8 @@ const createGoal = asyncHandler(async (req, res) => {
   }
 });
 
+
+
 // @desc Update Goals
 // @route Put /api/goals/id
 // @access Private
@@ -38,11 +44,24 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user.id)
+  // check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+  // ensure logged in user matched goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
   res.status(200).json(updatedGoal);
 });
+
+
 // @desc Delete Goals
 // @route Delete /api/goals/id
 // @access Private
@@ -52,8 +71,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not Found");
   }
+  const user = await User.findById(req.user.id)
+  // check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+  // ensure logged in user matched goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
   await Goal.findByIdAndRemove({ _id: req.params.id });
   res.status(200).json({ id: req.params.id });
 });
+
+
 
 module.exports = { getGoals, createGoal, updateGoal, deleteGoal };
